Module Tezos_client_016_PtMumbai.Managed_contract

val check_smart_contract : Protocol_client_context.full -> 'a option -> ('a -> 'b Lwt.t) -> 'b Lwt.t

check_smart_contract cctxt opt_value f_value returns an error if opt_value is None, and f_value value if opt_value = Some value. It can typically be used when opt_value is None for implicit accounts and Some _ for smart contracts, as the message in the raised error is that a smart contract is expected.

val get_contract_manager : Protocol_client_context.full -> Tezos_protocol_016_PtMumbai.Protocol.Contract_hash.t -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the manager key in a contract storage. The storage has to be of type `pair key_hash 'a`.

val build_delegate_operation : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?fee:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> Tezos_protocol_016_PtMumbai.Protocol.Contract_hash.t -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash option -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Kind.transaction Annotated_manager_operation.t Tezos_base.TzPervasives.tzresult Lwt.t

Builds a delegation operation ready for injection

val set_delegate : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?branch:int -> fee_parameter:Injection.fee_parameter -> ?fee:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> source:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> Tezos_protocol_016_PtMumbai.Protocol.Contract_hash.t -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash option -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Kind.transaction Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Kind.manager Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Set the delegate of a manageable contract. For a contract with a `do`entrypoint, it builds the lambda that set the provided delegate. `~source` has to be the registered manager of the contract.

val build_transaction_operation : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> contract:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Contract.t -> destination:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Contract.t -> ?entrypoint:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Entrypoint.t -> ?arg:string -> amount:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> ?fee:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> unit -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Kind.transaction Annotated_manager_operation.t Tezos_base.TzPervasives.tzresult Lwt.t

Builds a transaction operation ready for injection

val transfer : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?force:bool -> ?branch:int -> source:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys.sk_uri -> contract:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Contract.t -> destination:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Contract.t -> ?entrypoint:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Entrypoint.t -> ?arg:string -> amount:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> ?fee:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Manager_counter.t -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Kind.transaction Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Kind.manager Injection.result * Tezos_protocol_016_PtMumbai.Protocol.Contract_hash.t list) Tezos_base.TzPervasives.tzresult Lwt.t

Perform a transfer on behalf of a managed contract . For a contract with a `do`entrypoint, it builds the lambda that does the requested operation. `~source` has to be the registered manager of the contract.

val build_lambda_for_set_delegate : delegate: Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash option -> string

val build_lambda_for_transfer_to_implicit : destination: Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.public_key_hash -> amount:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> string

val build_lambda_for_transfer_to_originated : destination:Tezos_protocol_016_PtMumbai.Protocol.Contract_hash.t -> entrypoint:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Entrypoint.t -> amount:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Tez.t -> parameter_type:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Script.expr -> parameter:Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Script.expr -> string