Module Context_helpers.In_memory

In_memory is a context that can be used to instantiate an Arith or Wasm PVM. It's signature is Protocol.Alpha_context.Sc_rollup.ArithPVM.P = Protocol.Alpha_context.Sc_rollup.Wasm_2_0_0PVM.P

include P with type Tree.tree = Tezos_context_memory.Context_binary.tree and type Tree.t = Tezos_context_memory.Context_binary.t and type proof = Tezos_context_memory.Context.Proof.tree Tezos_context_memory.Context.Proof.t

module Tree : Tezos_context_sigs.Context.TREE with type key = string list and type value = bytes with type tree = Tezos_context_memory.Context_binary.tree with type t = Tezos_context_memory.Context_binary.t

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes. The trees are Merkle trees that carry the same hash as the part of the context they mirror.

type tree = Tree.tree

val hash_tree : tree -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Sc_rollup.State_hash.t

type proof = Tezos_context_memory.Context.Proof.tree Tezos_context_memory.Context.Proof.t

val proof_encoding : proof Tezos_base.TzPervasives.Data_encoding.t

val proof_before : proof -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Sc_rollup.State_hash.t

val proof_after : proof -> Tezos_protocol_016_PtMumbai.Protocol.Alpha_context.Sc_rollup.State_hash.t

val verify_proof : proof -> (tree -> (tree * 'a) Lwt.t) -> (tree * 'a) option Lwt.t

val produce_proof : Tree.t -> tree -> (tree -> (tree * 'a) Lwt.t) -> (proof * 'a) option Lwt.t

val make_empty_context : ?root:string -> unit -> Tree.t